requirements.txt
boto3
aio_pika
python-dotenv
azure-core
azure-ai-formrecognizer
asyncio
